[SYCL][E2E] Split bindless image read_sampled test
